refactor to prevent namespace trampling
